Refactored mixins
